Removed loaders from styleguidist dependencies, refactored Styleguidist's webpack, Capitalized Core and Default-Theme packages directories, indexes, and references
