+++ lerna.json
@@ -1,5 +1,5 @@
 {
-  "lerna": "2.0.0-rc.4",
+  "lerna": "2.0.0-rc.5",
   "version": "0.0.10",
   "packages": [
     "packages/*"
