+++ package.json
@@ -43,7 +43,6 @@
     "babel-plugin-transform-react-inline-elements": "^6.22.0",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.2",
     "colors": "^1.1.2",
-    "loaders": "^1.1.3",
     "react": "^15.5.4",
     "react-autosuggest": "^9.0.1",
     "react-docgen": "^2.14.1",
