+++ packages/react-atlas-core/src/Avatar/Avatar.js
