+++ packages/react-atlas-core/src/Avatar/index.js
