+++ packages/react-atlas-core/src/Checkbox/Checkbox.js
