+++ packages/react-atlas-core/src/Checkbox/index.js
