+++ packages/react-atlas-core/src/Switch/README.md
