+++ packages/react-atlas-core/src/Switch/index.js
