+++ packages/react-atlas-core/src/index.js
@@ -1,11 +1,11 @@
-export { AvatarCore } from "./avatar";
-export { ButtonCore } from "./button";
-export { FormCore } from "./form";
-export { InputCore } from "./input";
-export { RadioCore } from "./radio";
-export { RadioGroupCore } from "./radioGroup";
-export { TableCore } from "./table";
-export { SwitchCore } from "./switch";
-export { TooltipCore } from "./tooltip";
-export { CheckboxCore } from "./checkbox";
-export { DropdownCore } from "./dropdown";
+export { AvatarCore } from "./Avatar";
+export { ButtonCore } from "./Button";
+export { CheckboxCore } from "./Checkbox";
+export { DropdownCore } from "./Dropdown";
+export { FormCore } from "./Form";
+export { InputCore } from "./Input";
+export { RadioCore } from "./Radio";
+export { RadioGroupCore } from "./RadioGroup";
+export { SwitchCore } from "./Switch";
+export { TableCore } from "./Table";
+export { TooltipCore } from "./Tooltip";
