+++ packages/react-atlas-default-theme/src/Avatar/avatar.css
