+++ packages/react-atlas-default-theme/src/Avatar/index.js
