+++ packages/react-atlas-default-theme/src/Checkbox/index.js
