+++ packages/react-atlas-default-theme/src/Switch/index.js
