+++ packages/react-atlas-default-theme/src/Switch/switch.css
