+++ packages/react-atlas-default-theme/src/index.js
@@ -1,11 +1,11 @@
-export { AvatarStyle } from "./avatar";
-export { ButtonStyle } from "./button";
-export { CheckboxStyle } from "./checkbox";
-export { DropdownStyle } from "./dropdown";
-export { FormStyle } from "./form";
-export { InputStyle } from "./input";
-export { RadioStyle } from "./radio";
-export { RadioGroupStyle } from "./radioGroup";
-export { TableStyle } from "./table";
-export { SwitchStyle } from "./switch";
-export { TooltipStyle } from "./tooltip";
+export { AvatarStyle } from "./Avatar";
+export { ButtonStyle } from "./Button";
+export { CheckboxStyle } from "./Checkbox";
+export { DropdownStyle } from "./Dropdown";
+export { FormStyle } from "./Form";
+export { InputStyle } from "./Input";
+export { RadioStyle } from "./Radio";
+export { RadioGroupStyle } from "./RadioGroup";
+export { TableStyle } from "./Table";
+export { SwitchStyle } from "./Switch";
+export { TooltipStyle } from "./Tooltip";
